move plugin installer to an interface rebrand as an App Store
