+++ lib/index.html
@@ -23,7 +23,7 @@
 
         <ul class="list-group">
           <li class="list-group-item"><a href="/apps">Browser Apps</a></li>
-          <li class="list-group-item"><a href="/installplugins">Install Server Plugins</a></li>
+          <li class="list-group-item"><a href="/appstore">App Store</a></li>
           <li class="list-group-item"><a href="/plugins/configure">Server Plugins Configuration</a></li>
         </ul>
