+++ lib/interfaces/appstore.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
-    <title>Signal K Server Plugins</title>
+    <title>Signal K Server App Store</title>
 
     <!-- Bootstrap -->
     <link href="/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
@@ -19,7 +19,7 @@
   </head>
   <body>
     <div class="container">
-      <h1>Signal K Server Plugins</h1>
+      <h1>Signal K Server App Store</h1>
 
       <div/>
     </div>
